requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pydub
resemblyzer
scikit-learn
PyPDF2
python-docx
requests
aiofiles
python-multipart
